requirements.txt
streamlit
scikit-learn
python-dotenv
groq
